requirements.txt
progress-bar==8
PyQt5==5.15.7
requests==2.27.0
